+++ docs/changelog.rst
@@ -9,8 +9,12 @@ The changelog of the library code is presented below. **Breaking changes** were 
 
 -  Implemented an event-based camera
 -  Implemented an optical flow sensor
+-  Implemented a segmentation camera
 -  Implemented a thermal camera
--  Added a test of all camera types
+-  Extended look definition to support temperature maps
+-  Added water temperature
+-  Added air temperature, pressure, and humidity
+-  Added a test application for all camera types
 -  Updated marine snow rendering to use the same particle system for vision sensors attached to the same body
 
 1.4